Implements an AuditLog for the StudiyManager Backend ([Issue #417](https://github.com/MORE-Platform/more-studymanager-frontend/issues/417))
### AuditLog

The AuditLog includes:

* id - a unique id created by the DB
* created - the timestamp when the log was written to the database
* userId - the id of the authenticated user
* studyId - the id of the affected user
* Action - for now the full signature of the audited method
* State: `success`, `redirect`, `error` or `unknown`: Determined based on the method results. If applicable based on the http-status, `error` in case of an exception. `unknown` if it can not be determined
* timestamp - when the audited method completed
* resource - currently unused
* details: This includes additional information that are stored as JSON serialized string in the audit log
    * `param_<param-name>` additional parameters of the audited call. NOTE the maximum length of the JSON serialization of a parameter is limited via a config
    * `user_roles` - all roles assigned to the authenticated user for the given request
    * `header_content-type` the `content-type` http header
    * `header_location` the `location` http header

### Storage

The auditLog is stored in the MORE database in the table `audit_logs`

### Configuration

```yaml
  audit:
    # the study.status to audit (fallback if not set: all)
    study-states:
      - active
      - paused
      - closed
     details-byte-limit: 1000
```

The `study-states` define the states of studies where audit logging is active. If not set audit logs are active in all states. The `details-byte-limit` defines the maximum size of JSON String for parameter so that the serialization is included in the details of the audit log. This applies mainly to JSON parameters parsed in requests, but might also be triggered by long String values.

### Implementation

Auditing is implemented using Spring AoP (aspect oriented programming). The Aspect is bound to functionality (`JoinPoint`) using the `@Audited` Annotation. Only Methods with this annotation are considered for audit logging. The implementation assumes usage in `@RestController`, but is strictly not limited to it

Example Usage:

```java
    @Override
    @RequiresStudyRole(StudyRole.STUDY_ADMIN)
    @Audited
    public ResponseEntity<StudyDTO> setStatus(Long studyId, StatusChangeDTO statusChangeDTO) {
```
The affected `studyId` is extracted

* based on a `Long` parameter with the name `studyId`
* based on the return value being a `ResponseEntity<StudyDTO>` a `StudyDTO` or a `Study`

if the `studyId` can not be determined an Exception is thrown.