Prehab2Rehab #421: Part 2 - Cron to create OccurredObservations based on study participants timelines
* Corn runs all 5 minutes
* retrieves all active studies
* retrieves the timestamp of the latest `OccurredObservation` per study
* iterates over all participants of the study
* retrieves the Timeline for each participants
* creates `OccurredObservation` for all Observations that started after the latest and including the current time (minute resolution)

Changes:

* Introduces an `ACTIVE_STATES` EnumSet in the `Study.Status` enumeration. The two active states where hard coded in several locations in the code
* The StudyRepository & Service now allows to iterate over all Studies in a set of `Study.Status`.
    * returns a `Stream` to avoid loading all Studies into memory
    * expanded the component test for the StudyRepository to validate this functionality
* `UpsertOccurredObservationsCron`: Implements the business logic of creating `OccurredObservations` for study participants as described above

Possible Improvements:

* The period of 5 minutes is currently hard coded. Making this configureable is not much work.

Notes:

First tried to use the `SchedulingServie` and an `Job` Implementation to schedule a job for every active study, but gave up this idea as this infrastructure seamed complex. Especially storing the id returned by each submitted job seamed like a big burden, as it requires persistence, what is somehow provided by the SDK, but really hard to understand. Having different Cron Jobs for separate Studies running at an offset would be nice. But the advantage seamed not worth the effort.