#441 Refactoring of the Studymanager Backend that splits the Application Module in two parts:
1. the studymanager module (existing): now only contains things specific to the application like all the web services, authentication, collecting audit log events
2. the studymanager-services module (new): containing all the services, repositories, elastic search, push notifications, ...

The idea behind this change is that one can re-use all the service functionality in multiple applications. Already currently the gateway re-creates a lot of functionality already present in the studymanager backend. After this refactoring one could remove a lot of code from the gateway by reusing the studymanager-services module.

This refactoring was done in the context of UMM, because we plan to repackage and build upon functionality of the studymanager-services module.

TODOs: The resulting modules need to be deployed as Maven artifacts in the Guthub Maven Repository. For that the build pipeline needs to be extended